requirements.txt
kafka-python
dotenv
tweepy==3.9.0
pyspark==2.4.6
